+++ src/infra/src/10_network/99_locals.tf
@@ -1,7 +1,16 @@
 locals {
-  project          = "${var.prefix}-${var.env_short}-${var.location_short}"
-  project_nodomain = "${var.prefix}-${var.env_short}-${var.location_short}"
-  product          = "${var.prefix}-${var.env_short}"
+  project              = "${var.prefix}-${var.env_short}-${var.location_short}"
+  project_nodomain     = "${var.prefix}-${var.env_short}-${var.location_short}"
+  product              = "${var.prefix}-${var.env_short}"
+  tenants_file_path    = "${path.module}/../data/tenants/tenants.json"
+  raw_tenants          = jsondecode(file(local.tenants_file_path))
+  tenant_domain_prefix = var.env == "prod" ? "" : "${var.env}."
+  tenants = {
+    for tenant_key, tenant_data in local.raw_tenants : tenant_key => {
+      domain      = "${local.tenant_domain_prefix}${tenant_data.domain}"
+      admin_email = format("%s@%s", tenant_data.admin_mailbox, "${local.tenant_domain_prefix}${tenant_data.domain}")
+    }
+  }
 
   web_acl_rules = [
     {