+++ src/infra/src/70_domains/onemail_common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # infra-domains-onemail-common
 
+## 1.5.0
+
+### Minor Changes
+
+- fd7b3ac: Shared tenants json file and env vars for ECS and Lambda
+
+## 1.4.0
+
+### Minor Changes
+
+- 402d456: AWS SES Tenants configuration
+
 ## 1.3.0
 
 ### Minor Changes