+++ src/onemail/om-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # om-common
 
+## 1.0.4
+
+### Patch Changes
+
+- 545f2fa: logs
+
 ## 1.0.3
 
 ### Patch Changes